Changed directory structure and fixed paths accordingly.
