+++ package.json
@@ -13,7 +13,7 @@
     "nodeunit": "^0.10.2"
   },
   "scripts": {
-    "test": "nodeunit --reporter minimal test/"
+    "test": "nodeunit --reporter minimal test/unit/"
   },
   "gypfile": true
 }
